+++ BigDecimal/BigDecimal.csproj
@@ -45,12 +45,6 @@
 		</When>
 	</Choose>
 	<ItemGroup>
-	  <Compile Remove="BigDecimalNotPassing.cs" />
-	</ItemGroup>
-	<ItemGroup>
-	  <None Include="BigDecimalNotPassing.cs" />
-	</ItemGroup>
-	<ItemGroup>
 	  <PackageReference Include="Roslynator.Analyzers" Version="3.3.0">
 	    <PrivateAssets>all</PrivateAssets>
 	    <IncludeAssets>runtime; build; native; contentfiles; analyzers; buildtransitive</IncludeAssets>
